CertStore for Windows complete, base version nearly complete.
Windows cert store methods for establishing and revoking OS certificate
trust is complete. The BaseInMemoryCertificateStore object is almost
complete, but not quite. Methods are still needed to connect to an
upstream host and spoof its certificate on demand. Also, additional code
to free all the allocated contexts, certificates, keypairs and temporary
EC key objects per host are yet to be done.

TlsCapableHttpBridge is still mostly incomplete. In a previous iteration
of this proxy, I simply did a peek read from the client on connect and
used third party code to parse the SNI extension. I am now refactoring
to use openSSL provided callbacks. These callbacks complicate the issue
a great deal. Also, TlsCapableProxy could not continue until the memory
cert store object was complete, as the aforementioned design change is
imposed on the certificate store as well.