Initial OnTlsServerName implementation
Implemented OnTlsServerName template specialization for SslSocket type
in the TlsCapableHttpBridge source. This is untested, and errors still
need to be reported through a yet-to-be-implemeted interface, but aside
from that, barring any errors it is complete.

This is my first crack at using this callback, as I previously used code
taken from sni_proxy to simply do a peek read and parse the the hostname
before initiating any handshake with the client. Again, yet to be
tested.