TlsCapableHttpBridge implemented, untested
TlsCapableHttpBridge has had all members implemented, but is untested.
Almost surely there are bugs, even bad bugs, because Visual Studio code
completion, syntax highlighting and basically everything else useful
about a modern IDE completely dies when editing templated headers.

Added method(s) for manually overriding the payload of a request or
response. Can't believe I missed that the first time around.

EventReporter was basically scrapped and remade. It simply uses plain
predefined callbacks instead of any fancy and probably broken template
stuff.

HttpFilteringEngine was modified to not re-scan using typeless filters
when a response was supplied to the ShouldBlock() method. This is to put
into practice the optimization described in the header of not wastefully
checking a request twice.

In TlsCapableHttpBridge , the previous method of trying to use some
callbacks provided in openSSL to extract SNI hostname data was
completely destroyed and that functionality was rewritten. The bridge
now does a peek read on connected TLS clients and attempts to manually
parse SNI hostname data. Again, this is completely untested as of this
commit.

KnownHttpHeaders got SDHC related google compression headers added.
These are needed because such headers have be removed whenever
discovered to avoid SDHC being used and thus blocking us out of
inspecting payloads.

The C API is a disaster as of this commit, I was in the middle of
rewriting it and it's in a complete state of flux.

SslSocket type was renamed to TlsSocket to more accurately reflect the
intended purpose and constraints.

Adblock plus filtering options enum and bitset definition were moved to
a dedicated header to avoid inclusion of the entire base filter class
just to get those simple enums.

EngineCallbackTypes.h was created to provide a layer of separation
between the C API header and everything else that uses such callbacks.