All filtering is now external
Major overhaul, mostly deletions. The engine has been completely
transformed to simply externalize filtering function. The engine will
call two callbacks, giving the user 100% control over inspecting and
choosing to block http transactions at different stages of the
transaction.

This way, the method of filtering doesn't need to part of the proxy
itself. This is now simply a generic TLS proxy with an optional API to
inspect andblock the requests  and responses it processes.

Also fixed an issue with the Windows TCP table. The dwLocalPort members
of the tcp table structure can contain junk content in the upper bits,
making direct comparison risky. We now mask the lower 16 bits to ensure
this doesn't happen.