Integrate AnyCPU proxy
Added a new project that functions as an AnyCPU proxy to the arch-specific mixed-managed DLL.

To accomplish the AnyCPU proxy, we created a common interface package to be implemented by both the mixed managed DLL and the proxy DLL.

To do the above, we moved the common engine events and interface stuff out of the mixed managed DLL and into the common interface DLL project.

Added nuspec with custom targets file to ensure that mixed managed arch specific DLL's are staged in target project as well as any projects referencing the project that installs the package.

To avoid having to rebuild-all just to get fresh binaries moved between projects, created the FORCE-COPY empty project that copy fresh mixed-managed DLLs to the proxy AnyCPU projects.

Added a work around for a bug in http_parser.

Added preprocessor definition to force http_parser to be built in non-strict mode.

Fixed #117
Fixed #116